Static libraries project commit

+++ 0x09-static_libraries/0-memset.c
@@ -11,7 +11,7 @@
 * Return: a pointer to memory area @s
 */
 
-char *_memset(char *s, char b, size_t n)
+char *_memset(char *s, char b, unsigned int n)
 {
 	unsigned int mayor;
 	unsigned char *veev = s, value = b;
